requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.28
marshmallow==3.26.1
multidict==6.6.4
mypy-extensions==1.1.0
numpy==2.3.3
openai==1.107.3
orjson==3.11.3
packaging==25.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
pyyaml==6.0.2
redis==6.4.0
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
restrictedpython==8.0
sniffio==1.3.1
sqlalchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
yarl==1.20.1
zstandard==0.25.0
